Cache xkbcomp output
Construct a unique filename based on the display name and the RMLVO
values. If that file contains valid contents, use it. Otherwise,
compile the keymap to that file and don't unlink it so that it will be
re-used the next time the server runs.

v2: Only cache files that get written to XKM_OUTPUT_DIR.
    Add --enable-xkb-cache config option to enable this behaviour

Signed-off-by: Keith Packard <keithp@keithp.com>